pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
django = ">=5.1.7,<6.0.0"
daphne = ">=4.1.2,<5.0.0"
django-unfold = ">=0.22.0,<1.0.0"
transformers = ">=4.50.1,<5.0.0"
google-cloud-translate = ">=3.20.2,<4.0.0"
spacy = ">=3.8.4,<4.0.0"
jieba = ">=0.42.1,<0.43.0"
tqdm = ">=4.67.1,<5.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
dspy = ">=2.6.15,<3.0.0"
tantivy = ">=0.22.2,<0.23.0"
sentencepiece = ">=0.2.0,<0.3.0"
sentence-transformers = ">=3.4.1,<4.0.0"
litellm = ">=1.62.0,<2.0.0"
torch = { version = ">=2.6.0,<3.0.0", source = "pytorch_cpu", platform = "linux" }
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz" }
sacrebleu = "^2.6.0"
stanza = "^1.11.1"
nltk = "^3.9.4"
scikit-learn = "^1.8.0"
pymorphy3 = "^2.0.6"

[[tool.poetry.source]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
